requirements.txt
streamlit
llama-index
llama-index-llms-openai
llama-index-embeddings-openai
llama-index-readers-wikipedia
python-dotenv
nltk 